--- source/basic-redux/init/rootReducer.js
@@ -1,9 +1,0 @@
-import { combineReducers } from 'redux';
-
-import { postsReducer as posts } from '../bus/posts/reducer';
-import { uiReducer as ui } from '../../bus/ui/reducer';
-
-export const rootReducer = combineReducers({
-    posts,
-    ui,
-});
