+++ source/bus/init/rootSaga.js
@@ -1,5 +1,5 @@
 import { all, call } from 'redux-saga/effects';
-import { watchPosts }  from '../../basic-redux/bus/posts/saga/watchers';
+import { watchPosts }  from '../posts/saga/watchers';
 
 export function* rootSaga () {
     yield all([call(watchPosts)]);
