+++ source/bus/posts/actions.js
@@ -1,5 +1,5 @@
 import { types } from './types';
-import { api } from '../../../REST';
+import { api } from '../../REST/index';
 
 export const postsActions = {
     fillPosts: (posts) => {
