+++ source/bus/posts/saga/watchers.js
@@ -1,6 +1,6 @@
 import { takeEvery, all, call } from 'redux-saga/effects';
 import { types } from '../types';
-import { createPost } from './workers';
+import { createPost } from './workers/index';
 
 export function* watchCreatePost() {
     yield takeEvery(types.CREATE_POST_ASYNC, createPost)
