+++ source/components/Posts/index.js
@@ -12,7 +12,7 @@ import { mockedProfile } from '../../instruments/mockedData';
 // Components
 import { Composer, Catcher, Post } from '../../components';
 
-import { postsActions } from '../../basic-redux/bus/posts/actions'
+import { postsActions } from '../../bus/posts/actions'
 
 const mapStateToProps = (state) => {
     return {
